Fixed DMA Engine result type

+++ common/inc/hal_pwm.h
@@ -9,28 +9,28 @@ extern "C" {
 #include "ll_pwm.h"
 
 typedef struct {
-  uint pwmscale        : 4;
-  uint RESERVED        : 4;
-  uint pwmsticky       : 1;
-  uint pwmzerocmp      : 1;
-  uint pwmdeglitch     : 1;
-  uint RESERVED1       : 1;
-  uint pwmenalways     : 1;
-  uint pwmenoneshot    : 1;
-  uint RESERVED2       : 2;
-  uint pwmcmp0center   : 1;
-  uint pwmcmp1center   : 1;
-  uint pwmcmp2center   : 1;
-  uint pwmcmp3center   : 1;
-  uint RESERVED3       : 4;
-  uint pwmcmp0gang     : 1;
-  uint pwmcmp1gang     : 1;
-  uint pwmcmp2gang     : 1;
-  uint pwmcmp3gang     : 1;
-  uint pwmcmp0ip       : 1;
-  uint pwmcmp1ip       : 1;
-  uint pwmcmp2ip       : 1;
-  uint pwmcmp3ip       : 1;
+  uint8_t pwmscale        : 4;
+  uint8_t RESERVED        : 4;
+  uint8_t pwmsticky       : 1;
+  uint8_t pwmzerocmp      : 1;
+  uint8_t pwmdeglitch     : 1;
+  uint8_t RESERVED1       : 1;
+  uint8_t pwmenalways     : 1;
+  uint8_t pwmenoneshot    : 1;
+  uint8_t RESERVED2       : 2;
+  uint8_t pwmcmp0center   : 1;
+  uint8_t pwmcmp1center   : 1;
+  uint8_t pwmcmp2center   : 1;
+  uint8_t pwmcmp3center   : 1;
+  uint8_t RESERVED3       : 4;
+  uint8_t pwmcmp0gang     : 1;
+  uint8_t pwmcmp1gang     : 1;
+  uint8_t pwmcmp2gang     : 1;
+  uint8_t pwmcmp3gang     : 1;
+  uint8_t pwmcmp0ip       : 1;
+  uint8_t pwmcmp1ip       : 1;
+  uint8_t pwmcmp2ip       : 1;
+  uint8_t pwmcmp3ip       : 1;
 } PWM_InitTypeDef;
 
 static inline void HAL_PWM_enable(PWM_TypeDef *PWMx) {
